kafka-flink-postgres/kafka-producer/requirements.txt
kafka-python==2.0.2
schedule==1.1.0
Faker==15.1.3